requirements.txt
cffi==1.13.2
cryptography==2.8
img2pdf==0.3.3
importlib-metadata==1.4.0
jeepney==0.4.2
keyring==21.1.0
Pillow==6.2.1
pycparser==2.19
pyinsane2==2.0.13
PyQt5==5.13.2
PyQt5-sip==12.7.0
PyQt5-stubs==5.13.1.4
SecretStorage==3.1.2
six==1.14.0
SQLAlchemy==1.3.12
yagmail==0.11.224
zipp==2.1.0
